requirements.txt
requests
python-telegram-bot
pandas
pandas_ta
matplotlib
pytz
